Initial Database Implementation
I've already insetted the database in its entirety. There may be a need
to alter the current database I've put in place, but time will tell.

Currently the list view and entry views don't pull from/push to the
database. This is a little more tricky simply  because I need to work
directly with Thomas in order to get things in sync. I know what I need
to do in terms of fills and cursors, but not how he currently has the
list view created. I can't change anything in the entry view until the
list view is finished.