Gallery section now starts in proper place
The Gallery starts in the ID specific folder of the entry its currently
working on, in addition the image selected -should- be displayed in the
image view.  The action bar now says the right thing in relation to the
home screen.  Waiting on the database to finish the filling of the data
on the entry screen.